Enable event-based application of the preferences and remove save button
The values entered into the popup are now applied immediately after entering them. This means that clicking on the "enabled" checkbox will immediately toggle the state of the restriction. Furthermore, the changes in selected width can be seen directly.
The save button has been replaced with a button to close the popup.